handle path-related errors as a result of relocation, reorganize project/portfolio folders
